# api/src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"VenueModel": {
				"$ref": "#/components/schemas/DefaultSelection__36_VenuePayload_",
				"description": "Model Venue"
			},
			"Venue": {
				"$ref": "#/components/schemas/VenueModel",
				"description": "Model Venue"
			},
...
			"infer_typeofupdateVenueSchema_": {
				"properties": {
					"is_active": {
						"type": "boolean"
					},
					"address": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"UpdateVenueDto": {
				"$ref": "#/components/schemas/infer_typeofupdateVenueSchema_"
			},
...
			"SafeUser": {
				"$ref": "#/components/schemas/Omit_User.password_"
			},
...
			"UpdateUserDto": {
				"$ref": "#/components/schemas/infer_typeofupdateUserSchema_"
			},
...
			"TournamentRuleDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"tournament_id": {
						"type": "number",
						"format": "double"
					},
					"is_active": {
						"type": "boolean"
					},
					"points_per_win": {
						"type": "number",
						"format": "double"
					},
					"points_per_draw": {
						"type": "number",
						"format": "double"
					},
					"points_per_loss": {
						"type": "number",
						"format": "double"
					},
					"forfeit_score": {
						"type": "number",
						"format": "double"
					},
					"yellow_cards_suspension": {
						"type": "number",
						"format": "double"
					},
					"max_players_per_team": {
						"type": "number",
						"format": "double"
					},
					"min_players_per_team": {
						"type": "number",
						"format": "double"
					},
					"teams_advance_per_group": {
						"type": "number",
						"format": "double"
					},
					"tiebreaker_order": {
						"items": {
							"$ref": "#/components/schemas/TiebreakerOption"
						},
						"type": "array"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"user": {
						"properties": {
							"phone": {
								"type": "string",
								"nullable": true
							},
							"email": {
								"type": "string"
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"phone",
							"email",
							"name",
							"id"
						],
						"type": "object",
						"nullable": true
					},
					"tournament": {
						"properties": {
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object",
						"nullable": true
					}
				},
				"required": [
					"id",
					"tournament_id",
					"is_active",
					"points_per_win",
					"points_per_draw",
					"points_per_loss",
					"forfeit_score",
					"yellow_cards_suspension",
					"max_players_per_team",
					"min_players_per_team",
					"teams_advance_per_group",
					"tiebreaker_order",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UpdateTournamentRuleDto": {
				"$ref": "#/components/schemas/infer_typeofupdateTournamentRuleSchema_"
			},
...
			"Tournament": {
				"$ref": "#/components/schemas/TournamentModel",
				"description": "Model Tournament"
			},
...
			"UpdateTournamentDto": {
				"$ref": "#/components/schemas/infer_typeofupdateTournamentSchema_"
			},
...
			"Team": {
				"$ref": "#/components/schemas/TeamModel",
				"description": "Model Team"
			},
			"PaginatedResult_Team_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/Team"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/PaginationMeta"
					}
				},
				"required": [
					"data",
					"meta"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SeasonTeam": {
				"$ref": "#/components/schemas/SeasonTeamModel",
				"description": "Model SeasonTeam"
			},
			"PaginatedResult_SeasonTeam_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/SeasonTeam"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/PaginationMeta"
					}
				},
				"required": [
					"data",
					"meta"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SeasonTeamWithRelations": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"season_id": {
						"type": "number",
						"format": "double"
					},
					"team_id": {
						"type": "number",
						"format": "double"
					},
					"user_id": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"status": {
						"$ref": "#/components/schemas/SeasonTeamStatus"
					},
					"group_id": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"deleted_at": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"is_active": {
						"type": "boolean"
					},
					"season": {
						"properties": {
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object"
					},
					"team": {
						"properties": {
							"logo": {
								"type": "string",
								"nullable": true
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"logo",
							"name",
							"id"
						],
						"type": "object"
					},
					"user": {
						"properties": {
							"email": {
								"type": "string"
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"email",
							"name",
							"id"
						],
						"type": "object",
						"nullable": true
					},
					"group": {
						"properties": {
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object",
						"nullable": true
					}
				},
				"required": [
					"id",
					"season_id",
					"team_id",
					"user_id",
					"status",
					"group_id",
					"created_at",
					"updated_at",
					"deleted_at",
					"is_active",
					"season",
					"team",
					"user",
					"group"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SelfRegisterSeasonTeamDto": {
				"$ref": "#/components/schemas/infer_typeofselfRegisterSeasonTeamSchema_"
			},
...
			"AdminAddSeasonTeamDto": {
				"$ref": "#/components/schemas/infer_typeofadminAddSeasonTeamSchema_"
			},
...
			"UpdateSeasonTeamStatusDto": {
				"$ref": "#/components/schemas/infer_typeofupdateSeasonTeamStatusSchema_"
			},
...
			"AssignGroupDto": {
				"$ref": "#/components/schemas/infer_typeofassignGroupSchema_"
			},
...
			"Season": {
				"$ref": "#/components/schemas/SeasonModel",
				"description": "Model Season"
			},
			"PaginatedResult_Season_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/Season"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/PaginationMeta"
					}
				},
				"required": [
					"data",
					"meta"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CreateSeasonDto": {
				"$ref": "#/components/schemas/infer_typeofcreateSeasonSchema_"
			},
...
			"UpdateSeasonDto": {
				"$ref": "#/components/schemas/infer_typeofupdateSeasonSchema_"
			},
...
			"UpdateSeasonStatusDto": {
				"$ref": "#/components/schemas/infer_typeofUpdateSeasonStatusSchema_"
			},
...
			"Role": {
				"$ref": "#/components/schemas/RoleModel",
				"description": "Model Role"
			},
...
			"UpdateRoleDto": {
				"$ref": "#/components/schemas/infer_typeofupdateRoleSchema_"
			},
...
			"UpdatePlayerDto": {
				"$ref": "#/components/schemas/infer_typeofupdatePlayerSchema_"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/venues/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Venue"
								}
							}
						}
					}
				},
				"tags": [
					"Venues"
				],
				"security": [
					{
						"jwt": [
							"user",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Venue"
								}
							}
						}
					}
				},
				"tags": [
					"Venues"
				],
				"security": [
					{
						"jwt": [
							"user",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateVenueDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Venues"
				],
				"security": [
					{
						"jwt": [
							"user",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/users/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SafeUser"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SafeUser"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/tournamentrules/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TournamentRuleDto"
								}
							}
						}
					}
				},
				"tags": [
					"TournamentRules"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UpdateTournamentRuleDto"
								}
							}
						}
					}
				},
				"tags": [
					"TournamentRules"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTournamentRuleDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"TournamentRules"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/tournaments/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Tournament"
								}
							}
						}
					}
				},
				"tags": [
					"Tournaments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Tournament"
								}
							}
						}
					}
				},
				"tags": [
					"Tournaments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTournamentDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Tournaments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/teams": {
			"get": {
				"operationId": "FindAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResult_Team_"
								}
							}
						}
					}
				},
				"tags": [
					"Teams"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "per_page",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "q",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "direction",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Team"
								}
							}
						}
					}
				},
				"tags": [
					"Teams"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string"
									},
									"coach_name": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"logo": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"name"
								]
							}
						}
					}
				}
			}
		},
		"/teams/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Team"
								}
							}
						}
					}
				},
				"tags": [
					"Teams"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Team"
								}
							}
						}
					}
				},
				"tags": [
					"Teams"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string"
									},
									"coach_name": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"logo": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Teams"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/seasonteams": {
			"get": {
				"operationId": "FindAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResult_SeasonTeam_"
								}
							}
						}
					}
				},
				"tags": [
					"SeasonTeams"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "per_page",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "q",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "direction",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					}
				]
			},
			"post": {
				"operationId": "AdminAdd",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SeasonTeamWithRelations"
								}
							}
						}
					}
				},
				"description": "Admin add team trực tiếp",
				"tags": [
					"SeasonTeams"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminAddSeasonTeamDto"
							}
						}
					}
				}
			}
		},
		"/seasonteams/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SeasonTeam"
								}
							}
						}
					}
				},
				"tags": [
					"SeasonTeams"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"SeasonTeams"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/seasonteams/register": {
			"post": {
				"operationId": "SelfRegister",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SeasonTeamWithRelations"
								}
							}
						}
					}
				},
				"description": "Team leader tự đăng ký vào season",
				"tags": [
					"SeasonTeams"
				],
				"security": [
					{
						"jwt": [
							"leader"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SelfRegisterSeasonTeamDto"
							}
						}
					}
				}
			}
		},
		"/seasonteams/{id}/status": {
			"patch": {
				"operationId": "UpdateStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SeasonTeamWithRelations"
								}
							}
						}
					}
				},
				"tags": [
					"SeasonTeams"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateSeasonTeamStatusDto"
							}
						}
					}
				}
			}
		},
		"/seasonteams/{id}/group": {
			"patch": {
				"operationId": "AssignGroup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SeasonTeamWithRelations"
								}
							}
						}
					}
				},
				"description": "Assign team vào group sau draw",
				"tags": [
					"SeasonTeams"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AssignGroupDto"
							}
						}
					}
				}
			}
		},
		"/seasons": {
			"get": {
				"operationId": "FindAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResult_Season_"
								}
							}
						}
					}
				},
				"tags": [
					"Seasons"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "per_page",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "q",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "direction",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Season"
								}
							}
						}
					}
				},
				"tags": [
					"Seasons"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateSeasonDto"
							}
						}
					}
				}
			}
		},
		"/seasons/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Season"
								}
							}
						}
					}
				},
				"tags": [
					"Seasons"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Season"
								}
							}
						}
					}
				},
				"tags": [
					"Seasons"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateSeasonDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Seasons"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/seasons/{id}/status": {
			"patch": {
				"operationId": "UpdateStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Season"
								}
							}
						}
					}
				},
				"tags": [
					"Seasons"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateSeasonStatusDto"
							}
						}
					}
				}
			}
		},
...
		"/roles/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Role"
								}
							}
						}
					}
				},
				"tags": [
					"Roles"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Role"
								}
							}
						}
					}
				},
				"tags": [
					"Roles"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateRoleDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Roles"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/players/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PlayerDto"
								}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"organizing",
							"guest"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PlayerDto"
								}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"organizing",
							"guest"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePlayerDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"organizing",
							"guest"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}